+++ persistence-modules/spring-data-jpa-enterprise-2/README.md
@@ -3,7 +3,8 @@
 This module contains articles about Spring Data JPA used in enterprise applications such as transactions, sessions, naming conventions and more 
 
 ### Relevant Articles: 
-
+- [Custom Naming Convention with Spring Data JPA](https://www.baeldung.com/spring-data-jpa-custom-naming)
+- [Working with Lazy Element Collections in JPA](https://www.baeldung.com/java-jpa-lazy-collections)
 
 
 ### Eclipse Config
