+++ persistence-modules/spring-data-jpa-enterprise-2/src/main/java/com/baeldung/namingstrategy/Person.java
@@ -1,6 +1,5 @@
 package com.baeldung.namingstrategy;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
