+++ persistence-modules/spring-data-jpa-enterprise/README.md
@@ -7,8 +7,6 @@ This module contains articles about Spring Data JPA used in enterprise applicati
 - [Spring Data Java 8 Support](https://www.baeldung.com/spring-data-java-8)
 - [DB Integration Tests with Spring Boot and Testcontainers](https://www.baeldung.com/spring-boot-testcontainers-integration-test)
 - [A Guide to Spring’s Open Session in View](https://www.baeldung.com/spring-open-session-in-view)
-- [Working with Lazy Element Collections in JPA](https://www.baeldung.com/java-jpa-lazy-collections)
-- [Custom Naming Convention with Spring Data JPA](https://www.baeldung.com/spring-data-jpa-custom-naming)
 - [Partial Data Update With Spring Data](https://www.baeldung.com/spring-data-partial-update)
 - [Spring Data JPA @Modifying Annotation](https://www.baeldung.com/spring-data-jpa-modifying-annotation)
 - [Spring JPA – Multiple Databases](https://www.baeldung.com/spring-data-jpa-multiple-databases)
